requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.0.2
matplotlib==3.7.1
seaborn==0.12.2
joblib==1.1.0
openpyxl==3.1.2
xlrd==2.0.1
google-generativeai==0.3.1